Adding tests for the firm geocoding
