+++ spec/models/adviser_spec.rb
@@ -1,16 +1,8 @@
 RSpec.describe Adviser do
+  include QueueSpecHelper
+
   before do
     clear_job_queue
-  end
-
-  def clear_job_queue
-    ActiveJob::Base.queue_adapter.enqueued_jobs.clear
-  end
-
-  def queue_contains_a_job_for(job_class)
-    ActiveJob::Base.queue_adapter.enqueued_jobs.any? do |elem|
-      elem[:job] == job_class
-    end
   end
 
   describe '#geocoded?' do
